node/requirements.txt
# node/requirements.txt
Flask==2.2.2
Werkzeug==2.2.2
ecdsa==0.18.0
pycryptodome==3.20.0
rlp==3.0.0
requests==2.28.1
plyvel==1.5.0